reinvent:2020 implementation of AWS::GreengrassV2::ComponentVersion
This commit includes a fully-functional implementation of a CloudFormation
resource provider for Greengrass V2 ComponentVersion resources.

The name and version of the created ComponentVersion resource
are derived from the provided recipe source, whether that
is a Lambda function (where componentName & componentVers are are required parameters), or a recipe provided inline (where
name & version are supplied as part of the recipe content).

Given that ComponentVersion resources are immutable, only
the 'Tags' field on a ComponentVersion may be updated without
triggering replacement of the resource.

Incomplete implementation of AWS::GreengrassV2::Deployment removed.