Applied content of internal-only files from private-dev-reinvent-2020
In order to create a private branch based off of the squashed RC
branch prepared for GitHub, we need to: 1) merge the GitHub
history into Brazil history, and then 2) bring in internal-only
files that were previously on private-dev-reinvent-2020 (which
was mingled with non-squashed commits from public-dev-reinvent-2020).

Post-GA, we can start using `mainline` from here internally instead
of building from `private-dev-reinvent-2020`.