Explicitly specify a value for insertionOrder for the properties of type array
Uluru updated the schema validation rules and added checks to explicitly specify the insertionOrder for properties of type array. When set to true, the order in which array items are specified must be honored, and that changing the order of the array will indicate a change in the property.

cr: https://code.amazon.com/reviews/CR-61711410